requirements.txt
torch>=2.0.0
transformers>=4.40.0
peft>=0.11.0
datasets>=2.19.0
accelerate>=0.30.0
streamlit>=1.35.0
pydantic>=2.7.0
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.13
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
openai>=1.30.0
python-dotenv>=1.0.0
